requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
dynaconf==3.1.7
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
loguru==0.6.0
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.7
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
validators==0.18.2
whitenoise==6.0.0
win32-setctime==1.1.0
